chore: CopyBundleSettingsDialog TS convert
